+++ src/oracle/controller.js
@@ -7,7 +7,7 @@ import {
   queryRound,
   getLatestSubmittedRound,
 } from "../helpers/chain.js";
-import { MiddlewareENV } from '../helpers/middlewareEnv.js';
+import { MiddlewareENV } from '../helpers/MiddlewareEnv.js';
 import { logger } from "../helpers/logger.js";
 
 // Load environment variables
