+++ src/oracle/monitor.js
@@ -14,6 +14,7 @@ import { createServer } from "http";
 import { parse } from "url";
 import { iterateReverse } from "@agoric/casting";
 import { logger } from "../helpers/logger.js";
+import { MonitorENV } from "../helpers/MonitorEnv.js";
 
 
 let envvars = {};
